Add MDR Quagga and wireshark-gtk
Added the MDR version of Quagga so that MANET can be used. Added GTK version of wireshark.
The setcap thing failed on build for me so I have removed it. Wireshark works okay but gives warning as it runs as root.
The extra was failing I guess because the blank directory was not cloned in the repo. I have just commented out.